Refine on the same file that uses it
